refactor(robot-server): split apart experimental sessions router
