+++ robot-server/tests/sessions/router/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for /sessions routers."""
